Fixed some UI issues
-Removed curved corners for CardView on pre-Lollipop devices. It caused "padding".
-Dynamic size of HugDialog to make it work for all type of screens.
-Fixed Banner so it always wraps the image perfectly.
-Dynamic height of hug items.

+++ app/src/main/java/util/ImageLoader.java
@@ -39,7 +39,7 @@ public class ImageLoader extends AsyncTask<String, Void, Bitmap> {
         }
 
         // Scale bitmap to fit requested width. Keep aspect ratio
-        float scaleFactor = mRequestedWidth / bitmap.getWidth();
+        float scaleFactor = mRequestedWidth / (float)bitmap.getWidth();
         return Bitmap.createScaledBitmap(
                 bitmap, mRequestedWidth, (int) (bitmap.getHeight() * scaleFactor), true);
     }
